zlog: new version as an slog Handler
This implements an slog.Handler that defaults to JSON but seamlessly
upgrades to the native journald protocol.

It also automatically extracts pprof labels (if the context.Context is
provided) and OpenTelemetry baggage (if the context.Context is provided
and a filtering function is provided).

This needs some additional testing and documentation before merge+release.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>